fix di errore stupidissimo in Banca (i soldi venivano trasferiti al contrario -.-, fottuti prelievi e versamenti), problema di cast risolto (temporaneamnete), inserito controllo per evitare autopagamenti (che comunque non influenzano il capitale finale). DA TESTARE soprattutto perchè i giocatori non escono da prigione. Abbiamo implementato il rimanere dentro per tre turni?

+++ src/monopoly/Terreno.java
@@ -15,7 +15,7 @@ public class Terreno extends Proprieta {
 		this.colore = colore;
 	}
 	
-	public int calcolaAffitto(Giocatore g){
+	public double calcolaAffitto(Giocatore g){
 		return super.calcolaAffitto(g)*
 				((this.getProprietario().possiede(this.getCasella().getTabellone(), this.getColore()))?
 						2
